exchange.service.machine.learning/requirements.txt
absl-py==0.9.0
amqp==2.5.2
aniso8601==8.0.0
astor==0.8.1
astroid==2.4.0
astunparse==1.6.3
autopep8==1.5.2
billiard==3.6.3.0
cachetools==4.1.1
celery==4.4.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dnspython==1.16.0
eventlet==0.25.2
Flask==1.0.2
Flask-RESTful==0.3.6
gast==0.3.3
gevent==20.5.0
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graypy==2.1.0
greenlet==0.4.15
grpcio==1.28.1
h5py==2.10.0
idna==2.10
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.1
kombu==4.6.8
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
numpy==1.18.3
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.3
prompt-toolkit==1.0.14
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.6.1
PyInquirer==1.0.3
pylint==2.5.0
pymongo==3.10.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.4.4
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.16
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-hub==0.7.0
tensorflowjs==2.6.0
termcolor==1.1.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
